# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema Pesquisas — API Multi-Tenant",
        "description": "Backend REST API multiempresa com RBAC. Use POST /api/v1/auth/login para obter o JWT e clique em Authorize inserindo: Bearer {token}. Super Admins podem usar o header X-Empresa-Id para escopar requisições a uma empresa.",
        "contact": {
            "email": "admin@sistema.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realizar login",
                "description": "Autentica o usuário e retorna o token JWT.",
                "operationId": "5db09a35f965c6ca7529ed3c3cae8e21",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "senha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "superadmin@sistema.com"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Admin@2024"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string"
                                        },
                                        "dados": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "tipo": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                },
                                                "usuario": {
                                                    "$ref": "#/components/schemas/Usuario"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realizar logout",
                "description": "Invalida o token JWT adicionando-o à blacklist.",
                "operationId": "8f1a443d39c3bf383f851fbb51253c64",
                "responses": {
                    "200": {
                        "description": "Logout realizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Dados do usuário autenticado",
                "operationId": "1e5f382ba4fa0960d233875cab75a361",
                "responses": {
                    "200": {
                        "description": "Dados do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "$ref": "#/components/schemas/Usuario"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Renovar token JWT",
                "operationId": "05bd0482845348cafa4da9e4fb7d0ffb",
                "responses": {
                    "200": {
                        "description": "Token renovado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "dados": {
                                            "properties": {
                                                "token": {
                                                    "type": "string"
                                                },
                                                "tipo": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/trocar-senha": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Trocar senha do usuário autenticado",
                "operationId": "b829f582c5ddb0504a80e0291239a370",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "senha_atual",
                                    "nova_senha",
                                    "confirmacao"
                                ],
                                "properties": {
                                    "senha_atual": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "nova_senha": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "confirmacao": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/auth/recuperar-senha": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Solicitar recuperação de senha",
                "operationId": "711b75c591b92e8bc51192cfbf73c782",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token gerado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "sucesso": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "mensagem": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "E-mail não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RespostaErro"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RespostaErro": {
                "description": "Resposta de erro padrão da API",
                "properties": {
                    "sucesso": {
                        "type": "boolean",
                        "example": false
                    },
                    "mensagem": {
                        "type": "string",
                        "example": "Mensagem descritiva do erro."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token JWT obtido via POST /api/v1/auth/login. Insira: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}